third_party/blink/web_tests/external/wpt/css/selectors/invalidation/has-complexity-expected.txt

This is a testharness.js-based test.
PASS Before appending 25000 elements
FAIL After appending 25000 elements. This should not time out. assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL After appending another 25000 elements. This should not time out. assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
FAIL After appending div with 25000 elements. This should not time out. assert_equals: expected "rgb(128, 0, 128)" but got "rgb(255, 0, 0)"
FAIL After removing div with 25000 elements. This should not time out. assert_equals: expected "rgb(0, 0, 255)" but got "rgb(255, 0, 0)"
FAIL After removing 25000 elements one-by-one. This should not time out. assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL After removing the remaining elements. This should not time out. assert_equals: expected "rgb(128, 128, 128)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
